backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "casinoking-backend"
version = "0.1.0"
description = "CasinoKing backend bootstrap"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110,<1.0",
  "psycopg[binary]>=3.2,<4.0",
  "PyJWT>=2.8,<3.0",
  "uvicorn[standard]>=0.29,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8,<9",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
